feat: support config

+++ nas-music/src/main/java/top/itning/yunshunas/music/service/impl/SearchServiceImpl.java
@@ -5,8 +5,6 @@ import co.elastic.clients.elasticsearch._types.query_dsl.Query;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.data.elasticsearch.client.elc.ElasticsearchTemplate;
 import org.springframework.data.elasticsearch.client.elc.NativeQueryBuilder;
 import org.springframework.data.elasticsearch.core.SearchHits;
 import org.springframework.data.elasticsearch.core.query.HighlightQuery;
@@ -14,6 +12,7 @@ import org.springframework.data.elasticsearch.core.query.highlight.Highlight;
 import org.springframework.data.elasticsearch.core.query.highlight.HighlightField;
 import org.springframework.data.elasticsearch.core.query.highlight.HighlightFieldParameters;
 import org.springframework.stereotype.Service;
+import top.itning.yunshunas.music.config.ElasticsearchConfig;
 import top.itning.yunshunas.music.datasource.CoverDataSource;
 import top.itning.yunshunas.music.datasource.LyricDataSource;
 import top.itning.yunshunas.music.datasource.MusicDataSource;
@@ -37,7 +36,6 @@ import java.util.stream.Collectors;
  */
 @Slf4j
 @Service
-@ConditionalOnProperty(value = "spring.data.elasticsearch.repositories.enabled", havingValue = "true")
 public class SearchServiceImpl implements SearchService {
     /**
      * 歌词搜索字段
@@ -49,7 +47,7 @@ public class SearchServiceImpl implements SearchService {
     private final MusicDataSource musicDataSource;
     private final LyricDataSource lyricDataSource;
     private final CoverDataSource coverDataSource;
-    private final ElasticsearchTemplate elasticsearchTemplate;
+    private final ElasticsearchConfig elasticsearchConfig;
 
     @Autowired
     public SearchServiceImpl(LyricElasticsearchRepository lyricElasticsearchRepository,
@@ -57,17 +55,20 @@ public class SearchServiceImpl implements SearchService {
                              MusicDataSource musicDataSource,
                              LyricDataSource lyricDataSource,
                              CoverDataSource coverDataSource,
-                             ElasticsearchTemplate elasticsearchTemplate) {
+                             ElasticsearchConfig elasticsearchConfig) {
         this.lyricElasticsearchRepository = lyricElasticsearchRepository;
         this.musicRepository = musicRepository;
         this.musicDataSource = musicDataSource;
         this.lyricDataSource = lyricDataSource;
         this.coverDataSource = coverDataSource;
-        this.elasticsearchTemplate = elasticsearchTemplate;
+        this.elasticsearchConfig = elasticsearchConfig;
     }
 
     @Override
     public void saveOrUpdateLyric(String musicId, String lyricId, String content) {
+        if (!elasticsearchConfig.enabled()) {
+            return;
+        }
         if (StringUtils.isAnyBlank(musicId, lyricId)) {
             throw new IllegalArgumentException("音乐ID或歌词ID为空！");
         }
@@ -99,12 +100,18 @@ public class SearchServiceImpl implements SearchService {
 
     @Override
     public void deleteLyric(String lyricId) {
+        if (!elasticsearchConfig.enabled()) {
+            return;
+        }
         lyricElasticsearchRepository.deleteById(lyricId);
     }
 
     @Override
     public List<SearchResult> searchLyric(String keyword) {
-        SearchHits<Lyric> search = elasticsearchTemplate.search(
+        if (!elasticsearchConfig.enabled()) {
+            return Collections.emptyList();
+        }
+        SearchHits<Lyric> search = elasticsearchConfig.getElasticsearchTemplate().search(
                 new NativeQueryBuilder()
                         .withQuery(new Query.Builder().matchPhrase(new MatchPhraseQuery.Builder().field(SEARCH_FILED_FOR_LYRIC).query(keyword).build()).build())
                         .withHighlightQuery(new HighlightQuery(new Highlight(Collections.singletonList(new HighlightField(SEARCH_FILED_FOR_LYRIC, HighlightFieldParameters.builder().withFragmentSize(50).withNumberOfFragments(1).build()))), null))
